feat(core/highlight-vars): run in exported documents also
