+++ tests/unit/SpecHelper.js
@@ -20,7 +20,7 @@ export function makePluginDoc(
     "/src/core/dfn.js", // Needed for "plugins-done" event,
     ...plugins,
   ];
-  return getDoc(`
+  return getDoc(html`
     <!DOCTYPE html>
     <html lang="en">
       <head>
@@ -45,15 +45,20 @@ export function makePluginDoc(
                   value: { ready: Promise.reject(err) },
                 });
               }
+              await document.respec.ready;
             }
           }
           run(${JSON.stringify(plugins)});
         </script>
       </head>
-      <body>${body}</body>
+      <body>
+        ${body}
+      </body>
     </html>
   `);
 }
+
+export const html = String.raw;
 
 /**
  * @param {string} html
